+++ test/smokes/shellcheck/with_ignore/abc.bash
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+a=`foo`
+echo "$a"